requirements.txt
# Computer Vision & Image Processing
opencv-contrib-python>=4.8.0,<4.11.0  # Full version with contrib modules
mediapipe>=0.10.5                     # Face detection AI model (Apple Silicon compatible)

# macOS System Integration
pyobjc-framework-Quartz>=9.0          # Screen lock detection and management

# Optional development dependencies
pytest>=7.4.0                         # Testing framework
black>=23.0.0                         # Code formatting
flake8>=6.1.0                         # Code linting 

# Test dependencies
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0                # Async test support 